chore: added status table cleanup logic before backup with code refactor (#2542)
